JMM - adjustments to arc/nonarc via Nicolas Cellier, 64bit cleanup, fix zombie at quit time SQK-21
git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3481 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 497a526c74103a6c5dc8c05d00b6ed9c26f27ddd

+++ platforms/unix/vm/sqUnixHeartbeat.c
@@ -463,9 +463,9 @@ ioHeartbeatMilliseconds() { return beatMilliseconds; }
 unsigned long
 ioHeartbeatFrequency(int resetStats)
 {
-	unsigned duration = (ioUTCMicroseconds() - get64(frequencyMeasureStart))
+	unsigned long duration = (ioUTCMicroseconds() - get64(frequencyMeasureStart))
 						/ MicrosecondsPerSecond;
-	unsigned frequency = duration ? heartbeats / duration : 0;
+	unsigned long frequency = duration ? heartbeats / duration : 0;
 
 	if (resetStats) {
 		unsigned long long zero = 0;
